Large refactor:     - Flattened Action hierarchy so that all actions (i.e. all     individual possibilities of spell actions and feature actions, as     well as the standard attack/dodge/dash/disengage) are evaluated in     turn for priority.     - Introduced weight value for each action for priority sorting         - Ideally this weight will be dynamically altered by the 	evolving fight, but for now it is just hard-coded per action 	type per archetype     - Fixed some bugs where actions were attempted and failed with no     fallback action         - Example: distance to enemy calculated as in range for a REACH 	melee attack, but no valid cell in the Ring allowed for said 	reach attack         - Now Action::invoke returns a bool and the action loop restarts 	and skips that action if failure occurred     -Somewhat standardized (not fully) implementation of spells and     features as analogs of each other, where each has a tracker class     and those that entail actions have an action class separate from the     tracker

+++ include/Ring.h
@@ -50,7 +50,7 @@ public:
 	std::vector<Cell*> getNeighboringCells(Cell* c);
 	std::vector<Cell*> sortCellsBy(std::function<bool(const Cell*, const Cell*)> pred);
 	std::vector<Cell*> sortCellsByDistance(Cell* c);
-	std::vector<Cell*> cellsInRange(Cell* c, int range);
+	std::vector<Cell*> cellsInRange(Cell* c, float range);
 
 	void reset();
 private:
@@ -61,9 +61,9 @@ private:
 	Cell m_arena[RING_SIZE][RING_SIZE];
 };
 
-int dist(int x1, int y1, int x2, int y2);
-int dist(const Cell* a, int x, int y);
-int dist(const Cell* a, const Cell* b);
+float dist(int x1, int y1, int x2, int y2);
+float dist(const Cell* a, int x, int y);
+float dist(const Cell* a, const Cell* b);
 
 #endif // !KERF_RING_H
